test: separate xcm transfer tests
