+++ tests/system/tests/transfers/transfers.system.test.ts
@@ -6,22 +6,21 @@ import { AssetsPageElements } from '../../pages/_elements/AssetsPageElements';
 import { WatchOnlyAssetsPage } from '../../pages/assetsPage/WatchOnlyAssetsPage';
 import { test } from '../../utils/baseRegularFixture';
 import { getChainByName } from '../../utils/readConfig';
-import {
-  proxyTransferTestCase,
-  transferConstants,
-  transferTestCases,
-  xcmTransferTestCases,
-} from '../../utils/transferTestCases';
+import { proxyTransferTestCase, transferConstants, transferTestCases } from '../../utils/transferTestCases';
 
 const feature = 'Transfers';
 const story = 'Transfers tests';
 
+const setupTestMetadata = async (): Promise<void> => {
+  await allure.feature(feature);
+  await allure.story(story);
+  test.slow();
+};
+
 test.describe('Regular transfers', { tag: ['@regular-transfers', '@regress'] }, () => {
   for (const { chainName, assetId, amount, recipient } of transferTestCases) {
     test(`Multisig can make regular transfer on ${chainName}`, async ({ transfersPage }) => {
-      await allure.feature(feature);
-      await allure.story(story);
-      test.slow();
+      await setupTestMetadata();
 
       const walletModal = await transfersPage.openWalletManagement();
       await walletModal.searchAndSelectWallet(transferConstants.multisig_name);
@@ -42,9 +41,7 @@ test.describe('Regular transfers', { tag: ['@regular-transfers', '@regress'] }, 
 
   for (const { chainName, assetId, amount, recipient } of transferTestCases) {
     test(`Nova, single wallet, can make regular transfer on ${chainName}`, async ({ transfersPage }) => {
-      await allure.feature(feature);
-      await allure.story(story);
-      test.slow();
+      await setupTestMetadata();
 
       const walletModal = await transfersPage.openWalletManagement();
       await walletModal.searchAndSelectWallet(transferConstants.nova_name);
@@ -66,9 +63,7 @@ test.describe('Regular transfers', { tag: ['@regular-transfers', '@regress'] }, 
 
   for (const { chainName, assetId, amount, recipient } of proxyTransferTestCase) {
     test('Proxy wallet can make regular transfer', async ({ transfersPage }) => {
-      await allure.feature(feature);
-      await allure.story(story);
-      test.slow();
+      await setupTestMetadata();
 
       const walletModal = await transfersPage.openWalletManagement();
       await walletModal.searchAndSelectWallet(transferConstants.proxy_name);
@@ -90,9 +85,7 @@ test.describe('Regular transfers', { tag: ['@regular-transfers', '@regress'] }, 
 
   for (const { chainName, assetId, amount, recipient } of transferTestCases) {
     test(`Polkadot Vault, single wallet, can make regular transfer on ${chainName}`, async ({ transfersPage }) => {
-      await allure.feature(feature);
-      await allure.story(story);
-      test.slow();
+      await setupTestMetadata();
 
       const walletModal = await transfersPage.openWalletManagement();
       await walletModal.searchAndSelectWallet(transferConstants.vault_name);
@@ -113,8 +106,7 @@ test.describe('Regular transfers', { tag: ['@regular-transfers', '@regress'] }, 
   }
 
   test('Watch-only transfer buttons should not be visible', async ({ transfersPage }) => {
-    await allure.feature(feature);
-    await allure.story(story);
+    await setupTestMetadata();
 
     const walletModal = await transfersPage.openWalletManagement();
     await walletModal.searchAndSelectWallet(transferConstants.watch_only_name);
@@ -128,9 +120,7 @@ test.describe('Regular transfers', { tag: ['@regular-transfers', '@regress'] }, 
   test(`Should not be able to open transfer modal on ${transferConstants.watch_only_chain} in watch-only mode`, async ({
     transfersPage,
   }) => {
-    await allure.feature(feature);
-    await allure.story(story);
-    test.slow();
+    await setupTestMetadata();
 
     const walletModal = await transfersPage.openWalletManagement();
     await walletModal.searchAndSelectWallet(transferConstants.watch_only_name);
@@ -140,34 +130,4 @@ test.describe('Regular transfers', { tag: ['@regular-transfers', '@regress'] }, 
     const transferModal = await transfersPage.tryOpenTransfer(chain, transferConstants.watch_only_asset_id, 3000);
     expect(transferModal).toBeNull();
   });
-
-  for (const { chainName, assetId, xcmChainName, amount, recipient } of xcmTransferTestCases) {
-    // TODO: remove fail flag after the bug is fixed (#5327) AND add EVM chains for test
-    test.fail(
-      `Polkadot Vault, single wallet, can make regular xcm transfer from ${chainName} to ${xcmChainName}`,
-      async ({ transfersPage }) => {
-        await allure.feature(feature);
-        await allure.story(story);
-        test.slow();
-
-        const walletModal = await transfersPage.openWalletManagement();
-        await walletModal.searchAndSelectWallet(transferConstants.vault_name);
-
-        const chain = getChainByName(substrateChains, chainName);
-
-        const transferModal = await transfersPage.openTransfer(chain, assetId);
-        await transferModal.chooseXcmChain(xcmChainName);
-
-        await transferModal.fillRecipient(recipient);
-        await transferModal.expectTransferFeeNotZero();
-
-        await transferModal.fillAmount(amount);
-
-        const confirmationModal = await transferModal.openConfirmationModal();
-        const signingModal = await confirmationModal.confirm();
-
-        await signingModal.checkQRCode();
-      },
-    );
-  }
 });
